+++ packages/xchain-doge/src/const.ts
@@ -54,7 +54,7 @@ export const blockstreamExplorerProviders = {
 
 const testnetSochainProvider = new SochainProvider(
   'https://sochain.com/api/v3',
-  process.env.SOCHAIN_API_KEY || '',
+  'PLACEHOLDER_APIKEY',
   DOGEChain,
   AssetDOGE,
   8,
@@ -62,7 +62,7 @@ const testnetSochainProvider = new SochainProvider(
 )
 const mainnetSochainProvider = new SochainProvider(
   'https://sochain.com/api/v3',
-  process.env.SOCHAIN_API_KEY || '',
+  'PLACEHOLDER_APIKEY',
   DOGEChain,
   AssetDOGE,
   8,