+++ packages/xchain-litecoin/src/const.ts
@@ -52,7 +52,7 @@ export const explorerProviders = {
 
 const testnetSochainProvider = new SochainProvider(
   'https://sochain.com/api/v3',
-  process.env.SOCHAIN_API_KEY || '',
+  'PLACEHOLDER_APIKEY',
   LTCChain,
   AssetLTC,
   8,
@@ -60,7 +60,7 @@ const testnetSochainProvider = new SochainProvider(
 )
 const mainnetSochainProvider = new SochainProvider(
   'https://sochain.com/api/v3',
-  process.env.SOCHAIN_API_KEY || '',
+  'PLACEHOLDER_APIKEY',
   LTCChain,
   AssetLTC,
   8,