+++ packages/xchain-thorchain-amm/__e2e__/thorchain-doSwap-e2e.ts
@@ -33,7 +33,7 @@ const thorchainQueryStagenet = new ThorchainQuery(thorchainCacheStagenet)
 const mainnetWallet = new Wallet(process.env.MAINNETPHRASE || 'you forgot to set the phrase', thorchainQueryMainnet)
 const stagenetWallet = new Wallet(process.env.MAINNETPHRASE || 'you forgot to set the phrase', thorchainQueryStagenet)
 
-const mainetThorchainAmm = new ThorchainAMM(thorchainQueryMainnet)
+const mainnetThorchainAmm = new ThorchainAMM(thorchainQueryMainnet)
 const stagenetThorchainAmm = new ThorchainAMM(thorchainQueryStagenet)
 
 const sBTC = assetFromStringEx('BTC/BTC')
@@ -83,12 +83,14 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAsset: AssetRuneNative,
       destinationAddress: mainnetWallet.clients['THOR'].getAddress(),
       // affiliateFeePercent: 0.1,
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
 
-    const outPutCanSwap = await mainetThorchainAmm.estimateSwap(estimateSwapParams)
+    const outPutCanSwap = await mainnetThorchainAmm.estimateSwap(estimateSwapParams)
     print(outPutCanSwap)
-    const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
-    console.log(output)
+    const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+    //console.log(output)
     expect(output.hash).toBeTruthy()
   })
 
@@ -98,9 +100,11 @@ describe('xchain-swap doSwap Integration Tests', () => {
       input: new CryptoAmount(assetToBase(assetAmount('0.0001')), AssetBTC),
       destinationAsset: AssetRuneNative,
       destinationAddress: mainnetWallet.clients['BNB'].getAddress(), // Wrong chain to trigger error
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
     try {
-      const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+      const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
       console.log(output.hash)
       fail()
     } catch (error: any) {
@@ -117,8 +121,10 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAddress: mainnetWallet.clients['ETH'].getAddress(),
       slipLimit: new BigNumber(0.5),
       // affiliateFeePercent: 0.1,
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
-    const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+    const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
     console.log(output)
     expect(output.hash).toBeTruthy()
   })
@@ -130,8 +136,10 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAddress: mainnetWallet.clients['ETH'].getAddress(),
       slipLimit: new BigNumber(0.5),
       // affiliateFeePercent: 0.1,
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
-    const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+    const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
     console.log(output)
     expect(output.hash).toBeTruthy()
   })
@@ -143,17 +151,19 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAsset: sBTC,
       destinationAddress: mainnetWallet.clients[THORChain].getAddress(),
       slipLimit: new BigNumber(0.03),
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
     try {
       const outPutCanSwap = await thorchainQueryMainnet.estimateSwap(estimateSwapParams)
       print(outPutCanSwap)
       if (outPutCanSwap.txEstimate.canSwap) {
-        const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+        const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
         console.log(`Tx hash: ${output.hash},\n Tx url: ${output.url}\n WaitTime: ${output.waitTimeSeconds}`)
         expect(output).toBeTruthy()
       }
-    } catch (error: any) {
-      console.log(error.message)
+    } catch (error) {
+      console.error(error)
     }
   })
 
@@ -164,13 +174,15 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAsset: BUSD,
       destinationAddress: mainnetWallet.clients['BNB'].getAddress(),
       slipLimit: new BigNumber(0.09),
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
     try {
-      const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+      const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
       console.log(`tx hash: ${output.hash}, \n Tx url: ${output.url} \n WaitTime:${output.waitTimeSeconds}`)
       expect(output).toBeTruthy()
-    } catch (error: any) {
-      console.log(error.message)
+    } catch (error) {
+      console.error(error)
     }
   })
 
@@ -181,17 +193,19 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAsset: sETH,
       destinationAddress: mainnetWallet.clients['THOR'].getAddress(),
       slipLimit: new BigNumber(0.5),
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
     try {
       const outPutCanSwap = await thorchainQueryMainnet.estimateSwap(estimateSwapParams)
       console.log(JSON.stringify(outPutCanSwap))
       if (outPutCanSwap.txEstimate.canSwap) {
-        const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+        const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
         console.log(`Tx hash: ${output.hash},\n Tx url: ${output.url}\n WaitTime: ${output.waitTimeSeconds}`)
         expect(output).toBeTruthy()
       }
-    } catch (error: any) {
-      console.log(error.message)
+    } catch (error) {
+      console.error(error)
     }
   })
 
@@ -202,17 +216,19 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAsset: sETH,
       destinationAddress: mainnetWallet.clients['THOR'].getAddress(),
       slipLimit: new BigNumber(0.5),
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
     try {
       const outPutCanSwap = await thorchainQueryMainnet.estimateSwap(estimateSwapParams)
       console.log(JSON.stringify(outPutCanSwap))
       if (outPutCanSwap.txEstimate.canSwap) {
-        const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+        const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
         console.log(`Tx hash: ${output.hash},\n Tx url: ${output.url}\n WaitTime: ${output.waitTimeSeconds}`)
         expect(output).toBeTruthy()
       }
-    } catch (error: any) {
-      console.log(error.message)
+    } catch (error) {
+      console.error(error)
     }
   })
 
@@ -224,8 +240,10 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAddress: mainnetWallet.clients['BNB'].getAddress(),
       slipLimit: new BigNumber(0.5),
       // affiliateFeePercent: 0.1,
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
-    const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+    const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
     console.log(output)
     expect(output.hash).toBeTruthy()
   })
@@ -238,8 +256,10 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAddress: mainnetWallet.clients['ETH'].getAddress(),
       slipLimit: new BigNumber(0.5),
       // affiliateFeePercent: 0.1,
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
-    const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+    const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
     console.log(output)
     expect(output.hash).toBeTruthy()
   })
@@ -252,6 +272,8 @@ describe('xchain-swap doSwap Integration Tests', () => {
         destinationAsset: AssetRuneNative,
         destinationAddress: mainnetWallet.clients['THOR'].getAddress(),
         slipLimit: new BigNumber('0.5'),
+        wallet: mainnetWallet,
+        walletIndex: 0,
       }
       const ethHelper = new EthHelper(mainnetWallet.clients.ETH, thorchainQueryMainnet.thorchainCache)
       const approved = await ethHelper.isTCRouterApprovedToSpend(
@@ -263,10 +285,28 @@ describe('xchain-swap doSwap Integration Tests', () => {
         expect(result.hash).toBeTruthy()
         console.log(JSON.stringify(result, null, 2))
       } else {
-        const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+        const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
         console.log(output)
         expect(output.hash).toBeTruthy()
       }
+    } catch (error) {
+      console.error(error)
+    }
+  })
+
+  // From ERC to Rune - passes
+  it(`Should check validity of swap from ERC to Rune`, async () => {
+    try {
+      const estimateSwapParams = {
+        input: new CryptoAmount(assetToBase(assetAmount('0.005', USDT_DECIMAL)), USDT),
+        destinationAsset: AssetRuneNative,
+        destinationAddress: mainnetWallet.clients['THOR'].getAddress(),
+        slipLimit: new BigNumber('0.5'),
+        wallet: mainnetWallet,
+        walletIndex: 0,
+      }
+      const isValidSwap = await mainnetThorchainAmm.estimateSwap(estimateSwapParams)
+      console.log(isValidSwap)
     } catch (error) {
       console.error(error)
     }
@@ -278,6 +318,8 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAddress: mainnetWallet.clients['ETH'].getAddress(),
       destinationAsset: XRUNE,
       slipLimit: new BigNumber('0.5'),
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
     const ethHelper = new EthHelper(mainnetWallet.clients.ETH, thorchainQueryMainnet.thorchainCache)
     const approved = await ethHelper.isTCRouterApprovedToSpend(
@@ -288,7 +330,7 @@ describe('xchain-swap doSwap Integration Tests', () => {
       const result = await ethHelper.approveTCRouterToSpend(estimateSwapParams.input.asset)
       console.log(JSON.stringify(result, null, 2))
     }
-    const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+    const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
     console.log(output)
     expect(output.hash).toBeTruthy()
   })
@@ -298,6 +340,8 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAddress: stagenetWallet.clients[AVAXChain].getAddress(),
       destinationAsset: AssetAVAX,
       slipLimit: new BigNumber('0.5'),
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
     const output = await stagenetThorchainAmm.doSwap(stagenetWallet, estimateSwapParams)
     console.log(output)
@@ -309,8 +353,10 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAddress: mainnetWallet.clients['THOR'].getAddress(),
       destinationAsset: AssetRuneNative,
       slipLimit: new BigNumber('0.5'),
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
-    const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+    const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
     console.log(output)
     expect(output.hash).toBeTruthy()
   })
@@ -320,8 +366,10 @@ describe('xchain-swap doSwap Integration Tests', () => {
       destinationAddress: mainnetWallet.clients[THORChain].getAddress(),
       destinationAsset: sATOM,
       slipLimit: new BigNumber('0.05'),
+      wallet: mainnetWallet,
+      walletIndex: 0,
     }
-    const output = await mainetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
+    const output = await mainnetThorchainAmm.doSwap(mainnetWallet, estimateSwapParams)
     console.log(output)
     expect(output.hash).toBeTruthy()
   })
@@ -348,9 +396,8 @@ describe('xchain-swap doSwap Integration Tests', () => {
       //   console.log(`Tx hash: ${output.hash},\n Tx url: ${output.url}\n WaitTime: ${output.waitTimeSeconds}`)
       //   expect(output).toBeTruthy()
       // }
-    } catch (error: any) {
+    } catch (error) {
       console.error(error)
-      fail()
     }
   })
 })